Bug fix.
Fetching older items was wrongly getting items with id < given oldest id.
This was causing items that were older than the given oldest not to be selected
because their id's were higher even though their datetimes were older. Changed
to calculate the datetime of the item with the given oldest id and fetch items
older than it.